+++ Android~/src/APSPreBidCompletion.kt
@@ -1,7 +1,6 @@
-package com.chartboost.mediation.unity.aps
+package com.chartboost.mediation.unity.adapter.aps
 
 import com.chartboost.mediation.amazonpublisherservicesadapter.AmazonPublisherServicesAdapter
-
 interface APSPreBidCompletion {
     fun completion(preBidAdInfo: AmazonPublisherServicesAdapter.AmazonPublisherServicesAdapterPreBidAdInfo)
 }