+++ Runtime/IOS/AmazonPublisherServicesAdapter.cs
@@ -1,17 +1,18 @@
 using System;
 using System.Runtime.InteropServices;
 using AOT;
+using Chartboost.Constants;
+using Chartboost.Logging;
 using Chartboost.Mediation.AmazonPublisherServices.Common;
+using Chartboost.Mediation.Utilities;
 using UnityEngine;
 
 namespace Chartboost.Mediation.AmazonPublisherServices.IOS
 {
     #nullable enable
-    public class AmazonPublisherServicesAdapter : IAmazonPublisherServicesAdapter
+    internal sealed class AmazonPublisherServicesAdapter : IAmazonPublisherServicesAdapter
     {
-        private delegate void ChartboostMediationAmazonPublisherServicesPreBidConsumer(string chartboostPlacement, string adFormat, int height, int width, bool isVideo, string partnerPlacement);
-        
-        private const string Internal = "__Internal";
+        private delegate void ChartboostMediationAmazonPublisherServicesPreBidConsumer(string mediationPlacement, string adFormat, string keywordsJson, string bannerSizeJson, int height, int width, bool isVideo, string partnerPlacement);
         
         [RuntimeInitializeOnLoadMethod]
         private static void RegisterInstance()
@@ -20,13 +21,27 @@ namespace Chartboost.Mediation.AmazonPublisherServices.IOS
                 return;
             AmazonPublisherServices.AmazonPublisherServicesAdapter.Instance = new AmazonPublisherServicesAdapter();
         }
+
+        /// <inheritdoc/>
+        public string AdapterNativeVersion => _CBMAmazonPublisherServicesAdapterAdapterVersion();
         
+        /// <inheritdoc/>
+        public string PartnerSDKVersion => _CBMAmazonPublisherServicesAdapterPartnerSDKVersion();
+        
+        /// <inheritdoc/>
+        public string PartnerIdentifier => _CBMAmazonPublisherServicesAdapterPartnerId();
+        
+        /// <inheritdoc/>
+        public string PartnerDisplayName => _CBMAmazonPublisherServicesAdapterPartnerDisplayName();
+        
+        /// <inheritdoc/>
         public bool TestMode
         {
             get => _CBMAmazonPublisherServicesAdapterGetTestMode();
             set => _CBMAmazonPublisherServicesAdapterSetTestMode(value);
         }
         
+        /// <inheritdoc/>
         public bool VerboseLogging
         {
             get => _CBMAmazonPublisherServicesAdapterGetVerboseLogging();
@@ -35,6 +50,7 @@ namespace Chartboost.Mediation.AmazonPublisherServices.IOS
         
         private static PreBiddingListener? _preBiddingListenerStatic;
         
+        /// <inheritdoc/>
         public PreBiddingListener? PreBiddingListener
         {
             get => _preBiddingListenerStatic;
@@ -49,7 +65,7 @@ namespace Chartboost.Mediation.AmazonPublisherServices.IOS
         }
         
         [MonoPInvokeCallback(typeof(ChartboostMediationAmazonPublisherServicesPreBidConsumer))]
-        private static void OnPreBid(string chartboostPlacement, string adFormat, int height, int width, bool isVideo, string partnerPlacement)
+        private static void OnPreBid(string mediationPlacement, string adFormat, string keywordsJson, string bannerSizeJson, int height, int width, bool isVideo, string partnerPlacement)
         {
             if (_preBiddingListenerStatic == null)
                 return;
@@ -60,12 +76,12 @@ namespace Chartboost.Mediation.AmazonPublisherServices.IOS
                 try
                 {
                     var unityAmazonSettings = new AmazonSettings(partnerPlacement, isVideo, height, width);
-                    var unityPreBidRequest = new AmazonPublisherServicesAdapterPreBidRequest(chartboostPlacement, adFormat, unityAmazonSettings);
+                    var unityPreBidRequest = new AmazonPublisherServicesAdapterPreBidRequest(mediationPlacement, adFormat, keywordsJson.ToDictionary(), bannerSizeJson.ToBannerSize(), unityAmazonSettings);
                     unityPreBidAdInfo = await _preBiddingListenerStatic.OnPreBid(unityPreBidRequest);
                 }
-                catch (Exception e)
-                {
-                    Debug.LogError(e);
+                catch (Exception exception)
+                { 
+                    LogController.LogException(exception);
                 }
                 finally
                 {
@@ -73,14 +89,17 @@ namespace Chartboost.Mediation.AmazonPublisherServices.IOS
                 }
             });
         }
-
-        [DllImport(Internal)] private static extern void _CBMAmazonPublisherServicesSetupPreBiddingListener(ChartboostMediationAmazonPublisherServicesPreBidConsumer preBidConsumer);
-        [DllImport(Internal)] private static extern void _CBMAmazonPublisherServicesRemovePreBiddingListener();
-        [DllImport(Internal)] private static extern void _CBMAmazonPublisherServicesPreBidCompletion(string? pricePoint, string? bidInfo);
-        [DllImport(Internal)] private static extern bool _CBMAmazonPublisherServicesAdapterGetTestMode();
-        [DllImport(Internal)] private static extern void _CBMAmazonPublisherServicesAdapterSetTestMode(bool testMode); 
-        [DllImport(Internal)] private static extern bool _CBMAmazonPublisherServicesAdapterGetVerboseLogging();
-        [DllImport(Internal)] private static extern void _CBMAmazonPublisherServicesAdapterSetVerboseLogging(bool testMode);
+        
+        [DllImport(SharedIOSConstants.DLLImport)] private static extern string _CBMAmazonPublisherServicesAdapterAdapterVersion();
+        [DllImport(SharedIOSConstants.DLLImport)] private static extern string _CBMAmazonPublisherServicesAdapterPartnerSDKVersion();
+        [DllImport(SharedIOSConstants.DLLImport)] private static extern string _CBMAmazonPublisherServicesAdapterPartnerId();
+        [DllImport(SharedIOSConstants.DLLImport)] private static extern string _CBMAmazonPublisherServicesAdapterPartnerDisplayName();
+        [DllImport(SharedIOSConstants.DLLImport)] private static extern bool _CBMAmazonPublisherServicesAdapterGetTestMode();
+        [DllImport(SharedIOSConstants.DLLImport)] private static extern void _CBMAmazonPublisherServicesAdapterSetTestMode(bool testMode); 
+        [DllImport(SharedIOSConstants.DLLImport)] private static extern bool _CBMAmazonPublisherServicesAdapterGetVerboseLogging();
+        [DllImport(SharedIOSConstants.DLLImport)] private static extern void _CBMAmazonPublisherServicesAdapterSetVerboseLogging(bool testMode);
+        [DllImport(SharedIOSConstants.DLLImport)] private static extern void _CBMAmazonPublisherServicesSetupPreBiddingListener(ChartboostMediationAmazonPublisherServicesPreBidConsumer preBidConsumer);
+        [DllImport(SharedIOSConstants.DLLImport)] private static extern void _CBMAmazonPublisherServicesRemovePreBiddingListener();
+        [DllImport(SharedIOSConstants.DLLImport)] private static extern void _CBMAmazonPublisherServicesPreBidCompletion(string? pricePoint, string? bidInfo);
     }
-    #nullable disable
 }